move bit f-ns to ad_utils
