+++ src/admc/ad/ad_utils.h
@@ -63,4 +63,7 @@ QString dn_from_name_and_parent(const QString &name, const QString &parent, cons
 QString get_default_domain_from_krb5();
 QString domain_to_domain_dn(const QString &domain);
 
+int bit_set(int bitmask, int bit, bool set);
+bool bit_is_set(int bitmask, int bit);
+
 #endif /* AD_UTILS_H */
