+++ src/common/utils.h
@@ -46,9 +46,6 @@ QString get_dn_from_pos(const QPoint &pos, const QAbstractItemView *view, int dn
 
 QList<QStandardItem *> make_item_row(const int count);
 
-int bit_set(int bitmask, int bit, bool set);
-bool bit_is_set(int bitmask, int bit);
-
 void exec_menu_from_view(QMenu *menu, const QAbstractItemView *view, const QPoint &pos);
 // NOTE: view must have header items and model before this is called
 void setup_column_toggle_menu(const QTreeView *view, const QStandardItemModel *model, const QList<int> &initially_visible_columns);
